+++ PocketMon Game/animate.js
@@ -1,0 +1,224 @@
+import { initBattle } from './battleScene.js';
+import { background } from './modules/sprites/background.js';
+import { boundaries } from './modules/gameObjs/boundaries.js';
+import { battleZones} from './modules/gameObjs/boundaries.js';
+import { player } from './modules/sprites/player.js';
+import { foreground } from './modules/sprites/foreground.js';
+import { battle } from './modules/gameObjs/battle.js';
+import { keys } from './modules/gameObjs/keys.js';
+import { rectangularCollision } from './modules/gameObjs/rectangularCollision.js';
+import { audio } from './data/audio.js';
+import { animateBattle } from './battleScene.js';
+
+const movables = [background, ...boundaries, foreground, ...battleZones];
+
+let lastKey = '';
+
+window.addEventListener('keydown', (e) => {
+    switch (e.key) {
+        case 'w':
+            keys.w.pressed = true
+            lastKey = 'w'
+            break
+        case 'a':
+            keys.a.pressed = true
+            lastKey = 'a'
+            break
+        case 's':
+            keys.s.pressed = true
+            lastKey = 's'
+            break
+        case 'd':
+            keys.d.pressed = true
+            lastKey = 'd'
+            break
+    }
+});
+
+window.addEventListener('keyup', (e) => {
+    switch (e.key) {
+        case 'w':
+            keys.w.pressed = false
+            break
+        case 'a':
+            keys.a.pressed = false
+            break
+        case 's':
+            keys.s.pressed = false
+            break
+        case 'd':
+            keys.d.pressed = false
+            break
+    }
+});
+
+export function animate() {
+    const animationId = window.requestAnimationFrame(animate);
+    background.draw();
+    boundaries.forEach(boundary => {
+        boundary.draw();
+    })
+    battleZones.forEach(battleZone => {
+    })
+    player.draw();
+    foreground.draw();
+
+    let moving = true;
+    player.animate = false;
+
+    if (battle.initiated) return;
+
+    // activate a battle
+    if (keys.w.pressed || keys.a.pressed || keys.s.pressed || keys.d.pressed) {
+        for (let i = 0; i < battleZones.length; i++) {
+            const battleZone = battleZones[i]
+            const overlappingArea = (Math.min(player.position.x + player.width, battleZone.position.x + battleZone.width) - Math.max(player.position.x, battleZone.position.x)) * (Math.min(player.position.y + player.height, battleZone.position.y + battleZone.height) - Math.max(player.position.y, battleZone.position.y))
+            if (
+                rectangularCollision({
+                    rectangle1: player,
+                    rectangle2: battleZone
+                }) &&
+                overlappingArea > (player.width * player.height) / 2
+                && Math.random() < 0.01
+            ) {
+                //deactivate current animation loop
+                window.cancelAnimationFrame(animationId)
+
+                audio.Map.stop()
+                audio.initBattle.play()
+                audio.battle.play()
+                battle.initiated = true
+                gsap.to('#overlappingDiv', {
+                    opacity: 1,
+                    repeat: 3,
+                    yoyo: true,
+                    duration: 0.4,
+                    onComplete() {
+                        gsap.to('#overlappingDiv', {
+                            opacity: 1,
+                            duration: 0.4,
+                            onComplete() {
+                                initBattle()
+                                animateBattle()
+                                gsap.to('#overlappingDiv', {
+                                    opacity: 0,
+                                    duration: 0.4,
+                                })
+                            }
+                        })
+                    }
+                })
+                break;
+            }
+        }
+    }
+
+    if (keys.w.pressed && lastKey === 'w') {
+        player.animate = true
+        player.image = player.sprites.up
+
+        for (let i = 0; i < boundaries.length; i++) {
+            const boundary = boundaries[i]
+            if (
+                rectangularCollision({
+                    rectangle1: player,
+                    rectangle2: {
+                        ...boundary, position: {
+                            x: boundary.position.x,
+                            y: boundary.position.y + 3
+                        }
+                    }
+                })
+            ) {
+                // moving = false;
+                break;
+            }
+        }
+
+        if (moving)
+            movables.forEach(movable => {
+                movable.position.y += 3;
+            })
+    }
+    else if (keys.a.pressed && lastKey === 'a') {
+        player.animate = true;
+        player.image = player.sprites.left;
+
+        for (let i = 0; i < boundaries.length; i++) {
+            const boundary = boundaries[i];
+            if (
+                rectangularCollision({
+                    rectangle1: player,
+                    rectangle2: {
+                        ...boundary, position: {
+                            x: boundary.position.x + 3,
+                            y: boundary.position.y
+                        }
+                    }
+                })
+            ) {
+                moving = false
+                break
+            }
+        }
+
+        if (moving)
+            movables.forEach(movable => {
+                movable.position.x += 3;
+            })
+    }
+    else if (keys.s.pressed && lastKey === 's') {
+        player.animate = true;
+        player.image = player.sprites.down;
+
+        for (let i = 0; i < boundaries.length; i++) {
+            const boundary = boundaries[i];
+            if (
+                rectangularCollision({
+                    rectangle1: player,
+                    rectangle2: {
+                        ...boundary, position: {
+                            x: boundary.position.x,
+                            y: boundary.position.y - 3
+                        }
+                    }
+                })
+            ) {
+                moving = false
+                break
+            }
+        }
+
+        if (moving)
+            movables.forEach(movable => {
+                movable.position.y -= 3;
+            })
+    }
+    else if (keys.d.pressed && lastKey === 'd') {
+        player.animate = true;
+        player.image = player.sprites.right;
+
+        for (let i = 0; i < boundaries.length; i++) {
+            const boundary = boundaries[i];
+            if (
+                rectangularCollision({
+                    rectangle1: player,
+                    rectangle2: {
+                        ...boundary, position: {
+                            x: boundary.position.x - 3,
+                            y: boundary.position.y
+                        }
+                    }
+                })
+            ) {
+                moving = false;
+                break;
+            }
+        }
+
+        if (moving)
+            movables.forEach(movable => {
+                movable.position.x -= 3;
+            })
+    }
+}
