+++ PocketMon Game/canvas.js
@@ -1,0 +1,2 @@
+export const canvas = document.querySelector('canvas');
+export const c = canvas.getContext('2d');
