+++ PocketMon Game/classes.js
@@ -1,16 +1,18 @@
-class Sprite {
+import { c } from "./canvas.js";
+
+export class Sprite {
     constructor({ position, velocity, image, frames = { max: 1, hold: 10 }, sprites = [], animate = false, rotation = 0 }) {
-        this.position = position
-        this.image = new Image()
-        this.frames = { ...frames, val: 0, elapsed: 0 }
+        this.position = position;
+        this.image = new Image();
+        this.frames = { ...frames, val: 0, elapsed: 0 };
         this.image.onload = () => {
-            this.width = this.image.width / this.frames.max
-            this.height = this.image.height
+            this.width = this.image.width / this.frames.max;
+            this.height = this.image.height;
         }
-        this.image.src = image.src
-        this.animate = animate
-        this.sprites = sprites
-        this.opacity = 1
+        this.image.src = image.src;
+        this.animate = animate;
+        this.sprites = sprites;
+        this.opacity = 1;
     }
 
     draw() {
@@ -45,7 +47,7 @@ class Sprite {
     }
 }
 
-class Monster extends Sprite {
+export class Monster extends Sprite {
     constructor({
         position,
         velocity,
@@ -187,7 +189,7 @@ class Monster extends Sprite {
     }
 }
 
-class Boundary {
+export class Boundary {
     static width = 48
     static height = 48
     constructor({ position }) {
