+++ PocketMon Game/data/attacks.js
@@ -1,4 +1,4 @@
-const attacks = {
+export const attacks = {
     Tackle: {
         name: 'Tackle',
         damage: 10,
