+++ PocketMon Game/data/audio.js
@@ -1,4 +1,4 @@
-const audio = {
+export const audio = {
     Map: new Howl({
         src: './Audio/map.wav',
         html5: true,
