+++ PocketMon Game/index.html
@@ -136,11 +136,11 @@
 <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.9.1/gsap.min.js"
     integrity="sha512-H6cPm97FAsgIKmlBA4s774vqoN24V5gSQL4yBTDOY2su2DeXZVhQPxFK4P6GPdnZqM9fg1G3cMv5wD7e6cFLZQ=="
     crossorigin="anonymous" referrerpolicy="no-referrer"></script>
-<script src="./data/audio.js"></script>
+<!-- <script src="./data/audio.js"></script>
 <script src="./data/battleZones.js"></script>
 <script src="./data/collisions.js"></script>
 <script src="./data/attacks.js"></script>
 <script src="./monsters.js"></script>
-<script src="./classes.js"></script>
-<script src="./index.js"></script>
-<script src="./battleScene.js"></script>
+<script src="./classes.js"></script> -->
+<script type="module" src="./index.js"></script>
+<!-- <script src="./battleScene.js"></script> -->
