+++ PocketMon Game/modules/canvas/offset.js
@@ -1,0 +1,4 @@
+export const offset = {
+    x: -735,
+    y: -630
+}
