+++ PocketMon Game/modules/gameObjs/battle.js
@@ -1,0 +1,3 @@
+export const battle = {
+    initiated: false
+}
