+++ PocketMon Game/modules/gameObjs/keys.js
@@ -1,0 +1,14 @@
+export const keys = {
+    w: {
+        pressed: false
+    },
+    a: {
+        pressed: false
+    },
+    s: {
+        pressed: false
+    },
+    d: {
+        pressed: false
+    }
+};
