+++ PocketMon Game/modules/images/image.js
@@ -1,0 +1,2 @@
+export const image = new Image();
+image.src = './images/Pellet Town.png';
